Update diagram phase 2
The diagram of classes and templates of the previous phase didn´t look good. We have decided to do better for this phase (we decided to make the diagram in parts to see the classes and relationships well). For this phase we are asked to update the diagram, that is why we have changed the diagrams of the previous phase, to show in this only the updates.